requirements.txt
beautifulsoup4>=4.12.0
feedparser>=6.0.0
requests>=2.31.0
streamlit>=1.28.0
pandas>=1.5.0
aiohttp>=3.9.0
python-docx>=0.8.11
xlsxwriter>=3.0.0
lxml>=4.9.0